Created some camera magic for the agent.
The camera has an X and Y point, and a "vector point" it goes to. The
vector point is calculated by subtracting the camera from the object of
interest. Then the vector point is added onto the camera point to get the
"look point" when gluLookAt() is called. The camera points can be set
however needed, such as relative to the player or along some parametric
track.